Make anaglyphizing not a frame sink
We also implement a simple "most recent" frame sink to fill the void left by
some of the logic we ripped out of the anaglyphizer.
